refactor(create manifest): Relocate manifest pathing code to ArFSManifestToUpload PE-477
